--- a/blueocean-rest-impl/src/main/java/io/jenkins/blueocean/service/embedded/rest/AbstractPipelineImpl.java
+++ b/blueocean-rest-impl/src/main/java/io/jenkins/blueocean/service/embedded/rest/AbstractPipelineImpl.java
@@ -14,7 +14,6 @@ import hudson.model.Job;
 import hudson.model.Run;
 import hudson.model.User;
 import hudson.plugins.favorite.Favorites;
-import hudson.plugins.favorite.user.FavoriteUserProperty;
 import io.jenkins.blueocean.commons.ServiceException;
 import io.jenkins.blueocean.rest.Navigable;
 import io.jenkins.blueocean.rest.Reachable;
@@ -28,7 +27,6 @@ import io.jenkins.blueocean.rest.model.BlueQueueContainer;
 import io.jenkins.blueocean.rest.model.BlueRun;
 import io.jenkins.blueocean.rest.model.BlueRunContainer;
 import io.jenkins.blueocean.rest.model.Container;
-import io.jenkins.blueocean.rest.model.Containers;
 import io.jenkins.blueocean.rest.model.Resource;
 import io.jenkins.blueocean.service.embedded.util.FavoriteUtil;
 import org.kohsuke.stapler.Stapler;
@@ -41,6 +39,7 @@ import javax.annotation.Nullable;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
@@ -210,7 +209,58 @@ public class AbstractPipelineImpl extends BluePipeline {
 
     @Override
     public Container<Resource> getActivities() {
-        return Containers.fromResource(getLink(), Lists.newArrayList(Iterators.concat(getQueue().iterator(), getRuns().iterator())));
+
+        return new Container<Resource>(){
+            @Override
+            public Iterator<Resource> iterator() {
+                return Lists.newArrayList(Iterators.concat(getQueue().iterator(), getRuns().iterator())).iterator();
+            }
+
+            @Override
+            public Resource get(String name) {
+                return null; //not supported
+            }
+
+            @Override
+            public Link getLink() {
+                return AbstractPipelineImpl.this.getLink().rel("activities");
+            }
+
+            @Override
+            public Iterator<Resource> iterator(final int start, final int limit) {
+
+                return activityIterator(getQueue(), getRuns(), start, limit);
+            }
+        };
+    }
+
+    public static Iterator<Resource> activityIterator(final BlueQueueContainer queueContainer,
+                                                      final BlueRunContainer runContainer,
+                                                      final int start, final int limit){
+        final Iterator<? extends Resource> queueIterator = queueContainer.iterator();
+        int skipped = Iterators.skip(queueContainer.iterator(), start);
+        final Iterator<? extends Resource> runIterator = runContainer.iterator(start-skipped, limit);
+        return new Iterator<Resource>() {
+            int count=0;
+            @Override
+            public boolean hasNext() {
+                return count++ < limit &&(queueIterator.hasNext() || runIterator.hasNext());
+            }
+
+            @Override
+            public Resource next() {
+                if(queueIterator.hasNext()){
+                    return queueIterator.next();
+                }else{
+                    return runIterator.next();
+                }
+            }
+
+            @Override
+            public void remove() {
+                //noop
+            }
+        };
     }
 
     /**