--- a/blueocean-rest/src/main/java/io/jenkins/blueocean/rest/model/BluePipeline.java
+++ b/blueocean-rest/src/main/java/io/jenkins/blueocean/rest/model/BluePipeline.java
@@ -1,7 +1,5 @@
 package io.jenkins.blueocean.rest.model;
 
-import com.google.common.collect.Iterators;
-import com.google.common.collect.Lists;
 import io.jenkins.blueocean.commons.stapler.TreeResponse;
 import io.jenkins.blueocean.rest.Navigable;
 import io.jenkins.blueocean.rest.annotation.Capability;
@@ -126,9 +124,7 @@ public abstract class BluePipeline extends Resource {
      * @return Gives paginated concatenation of {#getQueue()} and {#getRuns()}, in that order
      */
     @Navigable
-    public Container<Resource> getActivities() {
-        return Containers.fromResource(getLink(), Lists.newArrayList(Iterators.concat(getQueue().iterator(), getRuns().iterator())));
-    }
+    public abstract Container<Resource> getActivities();
 
     @PUT
     @WebMethod(name="favorite")