@vivek 2016-11-17
@i386 @tfennelly @michaelneale besides code review it will be great if you guys can try it to see the perf improvement. Thanks.