@michaelneale 2016-11-18
test failures: 

```
getMultiBranchPipelineActivityRuns - io.jenkins.blueocean.rest.impl.pipeline.MultiBranchTest
a few seconds
Error
expected:<3> but was:<15>
Stacktrace
java.lang.AssertionError: expected:<3> but was:<15>
	at org.junit.Assert.fail(Assert.java:88)
```