@vivek 2016-11-18
@michaelneale @tfennelly @i386 @imeredith Implemented optimization of multi-branch pipelines /activities and /runs API. @tfennelly it will be great if you could test it.

My testing shows:

For 6000 pipeline jobs:
- First time load 7.1 sec to 212 ms
- reloads 625ms to 59ms

For multi-branch pipeline jobs (5 branches) with total about 1000+ jobs (most on master):
- First time load 2.3 sec to 437 ms
- reloads 86ms to 80 ms