@michaelneale 2016-11-18
Looks reasonable to me - quick tyre kick and nothing is on fire so... 

@tfennelly if this looks ok to you, and ATH passes etc - maybe worth a bee?