@michaelneale 2016-11-21
Looks like good progress so far, and like we discussed last week in terms of the algorithm. 
@kzantow spotted a bug, and @tfennelly was going to re-test it, so still a bit of a WIP I gather but we probably could get this in ASAP