@michaelneale 2016-11-22
had a discussion on this: this will change to calculate some MAX_ROWS set, which is then sorted and paginated server side (with a tuneable value). This means that pagination still makes sense and the page will work mostly as people expect but with a smaller amount of data (until we have a REAL DB we can query...)