@vivek 2016-11-22
@michaelneale @i386 code pushed. Please review.